Fixed interface implementation

+++ DistanceInterface.h
@@ -14,9 +14,9 @@ class DistanceInterface
 {
 public:
 	// This is the minimum range of the sensor
-	double minVal;
+	float minVal;
 	// This is the maximum range of the sensor
-	double maxVal;
+	float maxVal;
 
 	// Should return the CACHED value of the sensor
 	virtual float getDistance();
